+++ test/integration/068_partial_parsing_tests/extra-files/model_three_disabled2.sql
@@ -1,0 +1,13 @@
+- Disabled model
+{{ config(materialized='table', enabled=False) }}
+ 
+with source_data as (
+ 
+    select 1 as id
+    union all
+    select null as id
+ 
+)
+ 
+select *
+from source_data