+++ test/integration/068_partial_parsing_tests/extra-files/model_three_modified.sql
@@ -1,0 +1,14 @@
+{{ config(materialized='table') }}
+ 
+with source_data as (
+
+    {#- This is model three #}
+ 
+    select 1 as id
+    union all
+    select null as id
+ 
+)
+ 
+select *
+from source_data