+++ test/integration/068_partial_parsing_tests/extra-files/snapshot2.sql
@@ -1,0 +1,14 @@
+{% snapshot orders_snapshot %}
+
+{{
+    config(
+      target_schema=schema,
+      strategy='check',
+      unique_key='id',
+      check_cols=['status'],
+    )
+}}
+select * from {{ ref('orders') }}
+
+{% endsnapshot %}
+